reparacion de cambio de fase a listo, vista lic de laboratorio
