+++ modulo_laboratorio/views/OrdenExamenes.py
@@ -1,17 +1,10 @@
 #Django
-from django.http import JsonResponse, QueryDict
 from django.urls import reverse
-from dateutil.relativedelta import relativedelta
-from django.views.generic import View, TemplateView
-from django.shortcuts import redirect, render
-from django.db.utils import IntegrityError
-#Python 
-from datetime import datetime
-
+from django.views.generic import  TemplateView
+from django.shortcuts import redirect
 ##Libreria Propias
-from modulo_expediente.models import (Paciente)
-from modulo_laboratorio.models import (Categoria, EsperaExamen, Resultado)
-from modulo_laboratorio.serializers import Examenserializer, ResultadoSerializer
+from modulo_laboratorio.models import ( EsperaExamen)
+from modulo_laboratorio.views.Resultado import sync_cola
 
 class OrdenExamenCreate(TemplateView):
     
@@ -23,19 +16,8 @@ class OrdenExamenCreate(TemplateView):
             id_paciente=id_paciente
         )
         orden.save()
+        sync_cola()
         return redirect(reverse('update_orden_examenes',
                             kwargs={'id_paciente':id_paciente,'id_orden':orden.id},))
-    # def delete(self, request, *args, **kwargs):
-    #     id_consulta=int(self.kwargs['id_consulta'])
-    #     data = QueryDict(request.body)
-    #     OrdenExamenLaboratorio.objects.filter(
-    #         id_receta_orden_examen_laboratorio=data['id_receta'],
-    #         consulta_id=id_consulta
-    #     ).delete()
 
-    #     print(data)
-    #     self.response['type']='success'
-    #     self.response['data']='Receta de examenes eliminada.'
-
-    #     return JsonResponse(self.response)
