+++ templates/laboratorio/OrdenExamen/GestionOrdenExamenes.html
@@ -5,7 +5,7 @@
 
 {% endblock %}
 <!--Titulo-->
-{% block titulo %} Constancia Medica {% endblock %}
+{% block titulo %} Orden de Exámenes {% endblock %}
 {% block content %}
 <div class="container py-4">
 	<div class="row">
@@ -129,16 +129,7 @@
     let id_orden= '{{ id_orden }}';
     let socket_url=`${ws_scheme}://${window.location.host}/ws/laboratorio/cola/${tipo}/${id_orden}`;
     const colaLaboratorioSocket=new WebSocket(socket_url);
-    {% comment %} function sync_orden(){
-        colaLaboratorioSocket.send(JSON.stringify({
-                  'message':'sync',
-                  'tipo':tipo,
-                  'id_orden':
-                  }));
-    } {% endcomment %}
-    {% comment %} colaLaboratorioSocket.onopen =function() {
-        sync_orden();
-    } {% endcomment %}
+
     colaLaboratorioSocket.onmessage=(data)=>{
         let examenes=JSON.parse(data.data);
         puedeDescargarConsolidado=(examenes.puede_descargar==true)?'':'disabled';
@@ -166,7 +157,7 @@
             success: function(response) { 
 
                 render_mensajes(response.mensajes);
-                //sync_orden(); 
+
                 
             },
             error: function(response){
@@ -224,7 +215,7 @@
             }, 
             success: (data)=>{
                 render_mensajes(data.mensajes);
-                //sync_orden();
+
                 }
                 
             ,
